Document_Analysis/requirements.txt
fastapi
uvicorn
torch
transformers
sentencepiece
PyPDF2
pdf2image
pytesseract
Pillow